BAH-4659|Fix. Immunization history folder restructuring
