+++ apps/clinical/public/locales/locale_en.json
@@ -190,7 +190,6 @@
   "IMMUNIZATION_HISTORY_ADMINISTERED_ON_REQUIRED": "Please select the administered on date",
   "IMMUNIZATION_HISTORY_BATCH_NUMBER": "Batch Number",
   "IMMUNIZATION_HISTORY_BATCH_NUMBER_PLACEHOLDER": "Enter batch number",
-  "IMMUNIZATION_HISTORY_ENTER_BATCH_NUMBER": "Enter Batch Number",
   "IMMUNIZATION_HISTORY_BATCH_NUMBER_REQUIRED": "Please enter a batch number",
   "IMMUNIZATION_HISTORY_NO_BATCHES_AVAILABLE_ERROR": "No batch available",
   "IMMUNIZATION_HISTORY_DOSE_SEQUENCE": "Dose Sequence",
