+++ apps/clinical/src/components/forms/immunizationHistory/useBatchNumber.ts
@@ -5,8 +5,8 @@ import {
 } from '@bahmni/services';
 import { useQuery } from '@tanstack/react-query';
 import { useMemo, useState } from 'react';
-import { InputControlAttributes } from '../../../../../providers/clinicalConfig/models';
-import { getAvailableStocks } from '../../../../../services/inventoryService';
+import { InputControlAttributes } from '../../../providers/clinicalConfig/models';
+import { getAvailableStocks } from '../../../services/inventoryService';
 
 interface UseBatchNumberLogicProps {
   attributes: InputControlAttributes[] | undefined;
